+++ src/test/java/com/liberty52/product/ProductApplicationTests.java
@@ -2,10 +2,8 @@ package com.liberty52.product;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootTest
-@ComponentScan
 class ProductApplicationTests {
 
 	@Test
